Converted all components to classes; CSS isn't working
